LPS-91769 Move all server classes into one package
