+++ modules/test/arquillian-extension-junit-bridge/src/main/java/com/liferay/arquillian/extension/junit/bridge/server/ServerExecutorStatement.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.arquillian.extension.junit.bridge.statement;
+package com.liferay.arquillian.extension.junit.bridge.server;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.InvocationTargetException;
